Almost final Indexer.Fetcher.Optimism.DisputeGame
