+++ apps/indexer/lib/indexer/fetcher/optimism/output_root.ex
@@ -12,13 +12,14 @@ defmodule Indexer.Fetcher.Optimism.OutputRoot do
 
   import EthereumJSONRPC, only: [quantity_to_integer: 1]
 
+  alias Explorer.Application.Constants
   alias Explorer.{Chain, Helper, Repo}
-  alias Explorer.Chain.Optimism.OutputRoot
-  alias Indexer.Fetcher.Optimism.DisputeGame
+  alias Explorer.Chain.Optimism.{DisputeGame, OutputRoot}
   alias Indexer.Fetcher.{Optimism, RollupL1ReorgMonitor}
   alias Indexer.Helper, as: IndexerHelper
 
   @fetcher_name :optimism_output_roots
+  @stop_constant_key "optimism_output_roots_stopped"
 
   # 32-byte signature of the event OutputProposed(bytes32 indexed outputRoot, uint256 indexed l2OutputIndex, uint256 indexed l2BlockNumber, uint256 l1Timestamp)
   @output_proposed_event "0xa7aaf2512769da4e444e3de247be2564225c2e7a8f74cfe528e46e17d24868e2"
@@ -47,9 +48,13 @@ defmodule Indexer.Fetcher.Optimism.OutputRoot do
   def handle_continue(:ok, _state) do
     Logger.metadata(fetcher: @fetcher_name)
 
-    env = Application.get_all_env(:indexer)[__MODULE__]
-
-    Optimism.init_continue(env, env[:output_oracle], __MODULE__)
+    if Constants.get_constant_value(@stop_constant_key) == "true" do
+      Logger.warning("#{__MODULE__} will not start because dispute games exist.")
+      {:stop, :normal, %{}}
+    else
+      env = Application.get_all_env(:indexer)[__MODULE__]
+      Optimism.init_continue(env, env[:output_oracle], __MODULE__)
+    end
   end
 
   @impl GenServer
@@ -122,6 +127,7 @@ defmodule Indexer.Fetcher.Optimism.OutputRoot do
 
     if stop do
       Logger.warning("#{__MODULE__} is being stopped because dispute games exist.")
+      Constants.set_constant_value(@stop_constant_key, "true")
       {:stop, :normal, state}
     else
       new_start_block = last_written_block + 1
